+++ src/data/CreateData.js
@@ -1,0 +1,6 @@
+export const ENCRYPTION_OPTIONS = [
+  {value: "drm-public", label: "DRM - Public Access", title: "Playout Formats - Dash Widevine, HLS Sample AES, HLS AES-128"},
+  {value: "drm-all", label: "DRM - All Formats", title: "Playout Formats - Dash Widevine, HLS Sample AES, HLS AES-128, HLS Fairplay"},
+  {value: "drm-restricted", label: "DRM - Widevine and Fairplay", title: "Playout Formats - Dash Widevine, HLS Fairplay"},
+  {value: "clear", label: "Clear", title: "Playout Formats - HLS Clear"}
+];
