+++ src/data/HumanReadableText.js
@@ -1,0 +1,23 @@
+export const STATUS_TEXT = {
+  unconfigured: "Not Configured",
+  uninitialized: "Uninitialized",
+  inactive: "Inactive",
+  stopped: "Stopped",
+  starting: "Starting",
+  running: "Running",
+  stalled: "Stalled",
+  terminating: "Terminating"
+};
+
+export const FORMAT_TEXT = {
+  udp: "MPEGTS",
+  srt: "SRT",
+  "srt-caller": "SRT Caller",
+  rtmp: "RTMP"
+};
+
+export const CODEC_TEXT = {
+  h264: "H.264",
+  h265: "H.265",
+  mpeg2video: "MPEG-2"
+};
