+++ src/data/StreamData.js
@@ -1,0 +1,19 @@
+export const AV_STREAM = {
+  DEFAULT: "default",
+  CUSTOM: "specific"
+};
+
+export const ENCRYPTION_TYPES = {
+  DRM: "drm",
+  CLEAR: "clear"
+};
+
+export const STATUS_MAP = {
+  UNCONFIGURED: "unconfigured",
+  UNINITIALIZED: "uninitialized",
+  INACTIVE: "inactive",
+  STOPPED: "stopped",
+  STARTING: "starting",
+  RUNNING: "running",
+  STALLED: "stalled",
+};
