+++ src/index.html
@@ -6,7 +6,8 @@
     <title>Eluvio Live Stream</title>
     <meta content="text/html;charset=utf-8" http-equiv="Content-Type" />
     <meta name="apple-mobile-web-app-status-bar-style" content="black-translucent">
-    <script type="text/javascript" src="configuration.js"></script>
+    <script type="text/javascript" src="./configuration.js"></script>
+
   </head>
   <body>
     <div id="app" class="app"></div>
